services/telemetry_query/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.8.0
pyjwt==2.9.0
